node: Move `initialize` into a new `runner` module
