+++ node/src/main.rs
@@ -1,4 +1,6 @@
-mod initialize;
+//! Top-level node binary
+
+mod runner;
 mod options;
 
 #[tokio::main]
@@ -8,7 +10,7 @@ async fn main() -> anyhow::Result<()> {
     logging::init_logging(opts.log_path.as_ref());
     logging::log::trace!("Command line options: {:?}", opts);
 
-    let manager = initialize::initialize(opts).await?;
+    let manager = runner::initialize(opts).await?;
 
     #[allow(clippy::unit_arg)]
     Ok(manager.main().await)
